Move buffer length tracking from SocketInput to Connection - Required for connection filters, which use use two instances of SocketInput.  We need to ensure the combined buffers do not exceed the limit, and to coordinate the calls to `IConnectionControl` between the two instances of SocketInput.
