+++ src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/SocketInput.cs
@@ -18,6 +18,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
 
         private readonly MemoryPool _memory;
         private readonly IThreadPool _threadPool;
+        private readonly IBufferLengthControl _bufferLengthControl;
         private readonly ManualResetEventSlim _manualResetEvent = new ManualResetEventSlim(false, 0);
 
         private Action _awaitableState;
@@ -32,19 +33,12 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
         private bool _consuming;
         private bool _disposed;
 
-        private readonly BufferLengthConnectionController _bufferLengthConnectionController;
-
-        public SocketInput(MemoryPool memory, IThreadPool threadPool)
+        public SocketInput(MemoryPool memory, IThreadPool threadPool, IBufferLengthControl bufferLengthControl)
         {
             _memory = memory;
             _threadPool = threadPool;
+            _bufferLengthControl = bufferLengthControl;
             _awaitableState = _awaitableIsNotCompleted;
-        }
-
-        public SocketInput(MemoryPool memory, IThreadPool threadPool, int maxBufferLength, IConnectionControl connectionControl,
-             KestrelThread connectionThread) : this(memory, threadPool)
-        {
-            _bufferLengthConnectionController = new BufferLengthConnectionController(maxBufferLength, connectionControl, connectionThread);
         }
 
         public bool RemoteIntakeFin { get; set; }
@@ -72,7 +66,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
             lock (_sync)
             {
                 // Must call Add() before bytes are available to consumer, to ensure that Length is >= 0
-                _bufferLengthConnectionController?.Add(count);
+                _bufferLengthControl?.Add(count);
 
                 if (count > 0)
                 {
@@ -105,7 +99,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
             lock (_sync)
             {
                 // Must call Add() before bytes are available to consumer, to ensure that Length is >= 0
-                _bufferLengthConnectionController?.Add(count);
+                _bufferLengthControl?.Add(count);
 
                 if (_pinned != null)
                 {
@@ -210,9 +204,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
                         _head = consumed.Block;
                         _head.Start = consumed.Index;
 
-                        // Must call Subtract() after bytes have been freed, to avoid producer starting too early and growing
+                        // Must call Subtract() after _head has been advanced, to avoid producer starting too early and growing
                         // buffer beyond max length.
-                        _bufferLengthConnectionController?.Subtract(lengthConsumed);
+                        _bufferLengthControl?.Subtract(lengthConsumed);
                     }
 
                     if (!examined.IsDefault &&
@@ -341,73 +335,5 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
                 returnBlock.Pool.Return(returnBlock);
             }
         }
-
-        private class BufferLengthConnectionController
-        {
-            private readonly int _maxLength;
-            private readonly IConnectionControl _connectionControl;
-            private readonly KestrelThread _connectionThread;
-
-            private readonly object _lock = new object();
-
-            private int _length;
-            private bool _connectionPaused;
-
-            public BufferLengthConnectionController(int maxLength, IConnectionControl connectionControl, KestrelThread connectionThread)
-            {
-                _maxLength = maxLength;
-                _connectionControl = connectionControl;
-                _connectionThread = connectionThread;
-            }
-
-            public int Length
-            {
-                get
-                {
-                    return _length;
-                }
-                set
-                {
-                    // Caller should ensure that bytes are never consumed before the producer has called Add()
-                    Debug.Assert(value >= 0);
-
-                    _length = value;
-                }
-            }
-
-            public void Add(int count)
-            {
-                lock (_lock)
-                {
-                    // Add() should never be called while connection is paused, since ConnectionControl.Pause() runs on a libuv thread
-                    // and should take effect immediately.
-                    Debug.Assert(!_connectionPaused);
-
-                    Length += count;
-                    if (Length >= _maxLength)
-                    {
-                        _connectionPaused = true;
-                        _connectionControl.Pause();
-                    }
-                }
-            }
-
-            public void Subtract(int count)
-            {
-                lock (_lock)
-                {
-                    Length -= count;
-
-                    if (_connectionPaused && Length < _maxLength)
-                    {
-                        _connectionPaused = false;
-                        _connectionThread.Post(
-                            (connectionControl) => ((IConnectionControl)connectionControl).Resume(),
-                            _connectionControl);
-                    }
-                }
-            }
-        }
-
     }
 }
